Deduplicate code: extract ClosureDescriptorTrait, unify frontend branches
PHP:
- Extract identical describeClosure() from 3 adapters into shared
  Kernel\Inspector\ClosureDescriptorTrait (eliminates triple copy-paste)
- All three ConfigProviders now `use ClosureDescriptorTrait`

Frontend (EventsPage):
- Extract shared monoLinkSx style constant (was duplicated 5x)
- Unify two identical EventName branches (with/without member) into one
- Simplify isClosureDescriptor type guard (remove redundant `as any` cast)
