+++ libs/Adapter/Symfony/src/Inspector/SymfonyConfigProvider.php
@@ -4,9 +4,9 @@ declare(strict_types=1);
 
 namespace AppDevPanel\Adapter\Symfony\Inspector;
 
+use AppDevPanel\Kernel\Inspector\ClosureDescriptorTrait;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
-use Yiisoft\VarDumper\ClosureExporter;
 
 /**
  * Provides configuration data from Symfony's container for the ADP inspector.
@@ -16,7 +16,7 @@ use Yiisoft\VarDumper\ClosureExporter;
  */
 final class SymfonyConfigProvider
 {
-    private static ?ClosureExporter $closureExporter = null;
+    use ClosureDescriptorTrait;
 
     public function __construct(
         private readonly ContainerInterface $container,
@@ -139,20 +139,5 @@ final class SymfonyConfigProvider
             return $listener::class . '::__invoke';
         }
         return get_debug_type($listener);
-    }
-
-    /**
-     * @return array{__closure: true, source: string, file: string|null, startLine: int|null, endLine: int|null}
-     */
-    private static function describeClosure(\Closure $closure): array
-    {
-        $ref = new \ReflectionFunction($closure);
-        return [
-            '__closure' => true,
-            'source' => (self::$closureExporter ??= new ClosureExporter())->export($closure),
-            'file' => $ref->getFileName() ?: null,
-            'startLine' => $ref->getStartLine() ?: null,
-            'endLine' => $ref->getEndLine() ?: null,
-        ];
     }
 }
