+++ libs/frontend/packages/panel/src/Module/Inspector/Pages/EventsPage.tsx
@@ -45,7 +45,7 @@ function normalizeEntries(data: EventListenersType | null): EventEntry[] {
 }
 
 function isClosureDescriptor(value: unknown): value is ClosureDescriptor {
-    return typeof value === 'object' && value !== null && '__closure' in value && (value as any).__closure === true;
+    return typeof value === 'object' && value !== null && '__closure' in value;
 }
 
 function getListenerSearchText(listener: EventListener): string {
@@ -120,46 +120,29 @@ const StyledAccordionSummary = styled(AccordionSummary)(({theme}) => ({
 }));
 
 // ---------------------------------------------------------------------------
+// Shared sx constants
+// ---------------------------------------------------------------------------
+
+const monoLinkSx = {
+    fontFamily: primitives.fontFamilyMono,
+    fontSize: '12px',
+    color: 'primary.main',
+    wordBreak: 'break-all',
+    '&:hover': {textDecoration: 'underline'},
+} as const;
+
+// ---------------------------------------------------------------------------
 // Event name rendering
 // ---------------------------------------------------------------------------
 
 const EventName = React.memo(({name, eventClass}: {name: string; eventClass: string | null}) => {
     const parsed = parseEventName(name);
 
-    if (parsed && 'member' in parsed) {
-        return (
-            <FileLink className={parsed.className} methodName={parsed.member} sx={{minWidth: 0}}>
-                <Typography
-                    component="span"
-                    sx={{
-                        fontFamily: primitives.fontFamilyMono,
-                        fontWeight: 600,
-                        fontSize: '13px',
-                        color: 'primary.main',
-                        wordBreak: 'break-all',
-                        '&:hover': {textDecoration: 'underline'},
-                    }}
-                >
-                    {name}
-                </Typography>
-            </FileLink>
-        );
-    }
-
     if (parsed) {
+        const methodName = 'member' in parsed ? parsed.member : undefined;
         return (
-            <FileLink className={parsed.className} sx={{minWidth: 0}}>
-                <Typography
-                    component="span"
-                    sx={{
-                        fontFamily: primitives.fontFamilyMono,
-                        fontWeight: 600,
-                        fontSize: '13px',
-                        color: 'primary.main',
-                        wordBreak: 'break-all',
-                        '&:hover': {textDecoration: 'underline'},
-                    }}
-                >
+            <FileLink className={parsed.className} methodName={methodName} sx={{minWidth: 0}}>
+                <Typography component="span" sx={{...monoLinkSx, fontWeight: 600, fontSize: '13px'}}>
                     {name}
                 </Typography>
             </FileLink>
@@ -237,16 +220,7 @@ const ListenerItem = React.memo(({listener}: {listener: EventListener}) => {
         return (
             <ListenerRow>
                 <FileLink className={parsed.className} methodName={parsed.methodName} sx={{flex: 1, minWidth: 0}}>
-                    <Typography
-                        component="span"
-                        sx={{
-                            fontFamily: primitives.fontFamilyMono,
-                            fontSize: '12px',
-                            color: 'primary.main',
-                            wordBreak: 'break-all',
-                            '&:hover': {textDecoration: 'underline'},
-                        }}
-                    >
+                    <Typography component="span" sx={monoLinkSx}>
                         {serializeCallable(listener)}
                     </Typography>
                 </FileLink>
@@ -258,16 +232,7 @@ const ListenerItem = React.memo(({listener}: {listener: EventListener}) => {
         return (
             <ListenerRow>
                 <FileLink className={listener} sx={{flex: 1, minWidth: 0}}>
-                    <Typography
-                        component="span"
-                        sx={{
-                            fontFamily: primitives.fontFamilyMono,
-                            fontSize: '12px',
-                            color: 'primary.main',
-                            wordBreak: 'break-all',
-                            '&:hover': {textDecoration: 'underline'},
-                        }}
-                    >
+                    <Typography component="span" sx={monoLinkSx}>
                         {listener}
                     </Typography>
                 </FileLink>
@@ -277,16 +242,7 @@ const ListenerItem = React.memo(({listener}: {listener: EventListener}) => {
 
     return (
         <ListenerRow>
-            <Typography
-                sx={{
-                    flex: 1,
-                    minWidth: 0,
-                    fontFamily: primitives.fontFamilyMono,
-                    fontSize: '12px',
-                    color: 'text.secondary',
-                    wordBreak: 'break-all',
-                }}
-            >
+            <Typography sx={{...monoLinkSx, flex: 1, color: 'text.secondary'}}>
                 {serializeCallable(listener)}
             </Typography>
         </ListenerRow>
